requirements-comfyui.txt
accelerate==0.33.0
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
boto3==1.34.11
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
chardet==4.0.0
clip-interrogator==0.6.0
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==43.0.0
cupy-cuda12x==12.3.0
cupy-wheel==12.3.0
cycler==0.12.1
Deprecated==1.2.14
diffusers==0.30.0
einops==0.8.0
fastrlock==0.8.2
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==2.10
imageio==2.35.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
kornia>=0.7.1
kornia_rs==0.1.5
lark-parser==0.12.0
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.8.0
matrix-client==0.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.19.0
open_clip_torch==2.26.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
piexif==1.1.3
Pillow==10.4.0
protobuf==5.27.3
psutil==6.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.8.2
PyGithub==2.3.0
PyJWT==2.9.0
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.24
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.25.1
safetensors>=0.3.0
scikit-image==0.24.0
scipy==1.14.0
seaborn==0.13.2
segment-anything==1.0
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
spandrel==0.3.4
sympy==1.13.2
tifffile==2024.8.10
timm==1.0.8
tokenizers==0.19.1
torch
torchsde
torchvision
tqdm==4.66.4
trampoline==0.1.2
transformers>=4.25.1
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.78
ultralytics-thop==2.0.0
urllib3==1.26.19
wcwidth==0.2.13
websocket-client==1.8.0
wrapt==1.16.0
xformers==0.0.27.post2
yarl==1.9.4
zipp==3.20.0
toml==0.10.2
loguru==0.7.2
